# Optimize Uuid.ToString() by avoiding using sprintf to do the basic string transformation