--- a/sdk/core/azure-core/src/uuid.cpp
+++ b/sdk/core/azure-core/src/uuid.cpp
@@ -16,40 +16,35 @@ static thread_local std::mt19937_64 randomGenerator(std::random_device{}());
 } // namespace
 #endif
 
+namespace {
+static char ByteToHexChar(uint8_t byte)
+{
+  if (byte <= 9)
+    return '0' + byte;
+  if (byte >= 10 && byte <= 15)
+    return 'a' + (byte - 10);
+  throw std::out_of_range("Byte value out of range for hexadecimal character.");
+}
+} // namespace
+
 namespace Azure { namespace Core {
-  std::string Uuid::ToString()
+  std::string Uuid::ToString() const
   {
-    if (!m_uuidString.empty())
+    std::string s(36, '-');
+
+    for (size_t i = 0, j = 0; j < s.size() && i < UuidSize; i++)
     {
-      return m_uuidString;
+      if (i == 4 || i == 6 || i == 8 || i == 10)
+      {
+        j++; // Add hyphens at the appropriate places
+      }
+
+      uint8_t highNibble = (m_uuid[i] >> 4) & 0x0F;
+      uint8_t lowNibble = m_uuid[i] & 0x0F;
+      s[j++] = ByteToHexChar(highNibble);
+      s[j++] = ByteToHexChar(lowNibble);
     }
-    // Guid is 36 characters
-    //  Add one byte for the \0
-    char s[37];
-
-    std::snprintf(
-        s,
-        sizeof(s),
-        "%2.2x%2.2x%2.2x%2.2x-%2.2x%2.2x-%2.2x%2.2x-%2.2x%2.2x-%2.2x%2.2x%2.2x%2.2x%2.2x%2.2x",
-        m_uuid[0],
-        m_uuid[1],
-        m_uuid[2],
-        m_uuid[3],
-        m_uuid[4],
-        m_uuid[5],
-        m_uuid[6],
-        m_uuid[7],
-        m_uuid[8],
-        m_uuid[9],
-        m_uuid[10],
-        m_uuid[11],
-        m_uuid[12],
-        m_uuid[13],
-        m_uuid[14],
-        m_uuid[15]);
-
-    m_uuidString = std::string(s);
-    return m_uuidString;
+    return s;
   }
 
   Uuid Uuid::CreateUuid()