--- a/sdk/core/azure-core/test/ut/uuid_test.cpp
+++ b/sdk/core/azure-core/test/ut/uuid_test.cpp
@@ -16,6 +16,20 @@ TEST(Uuid, Basic)
   EXPECT_EQ(uuid.ToString().length(), 36);
 }
 
+TEST(Uuid, Roundtrip)
+{
+  std::array<uint8_t, 16U> uuidArray
+      = {97, 126, 195, 45, 41, 178, 70, 23, 142, 131, 221, 245, 20, 45, 215, 15};
+
+  auto uuid = Uuid::CreateFromArray(uuidArray);
+  std::string uuidString = uuid.ToString();
+  std::string expectedString = "617ec32d-29b2-4617-8e83-ddf5142dd70f";
+  EXPECT_EQ(expectedString, uuidString);
+
+  auto roundTrip = uuid.AsArray();
+  EXPECT_EQ(uuidArray, roundTrip);
+}
+
 TEST(Uuid, Transparent)
 {
   auto uuid1 = Uuid::CreateUuid();