Updated to measure the performance of the following:
```C++
auto uuidString = Azure::Core::Uuid::CreateUuid().ToString();
```
**Before:**
Completed 279,236 operations in a weighted-average of 10s (27,913.582653 ops/s, 3.58249e-05 s/op)

**After:**
Completed 920,677 operations in a weighted-average of 10s (91,995.810794 ops/s, 1.08701e-05 s/op)

New implementation results in **3.3x improvement**

Inspired by the conversation here: https://github.com/Azure/azure-sdk-for-cpp/pull/5914#discussion_r1718125489