@LarryOsterman 2024-08-20
I like the current iteration of these changes, IMHO they're a reasonable example of the kind of optimization that is almost certainly going to result in improvements, regardless of measurements.