@antkmsft 2024-08-23
I have a PR in progress, which adds `Parse()`, default constructor, and comparison operators. I am merging this PR, and will rebase my changes on top of this, so there are no merge conflicts between the two.